The leaderboard class is revised and Some parts of University class is changed
The ui elements are transferred to the leaderboardScreen class. The sorting mechanism is changed to a simple one, and the mechanism of the reputation point is implemented, which is directly affected by the student satisfaction level( this % might be changed later on). Also, in the university class, compareTo and toString methods are implemented with the constructor of the university class.